Refactor the Block component's detectVideoId function into a utility that can be shared by Block, Gallery & Header, and extend it to detect PL video IDs
